Rename ImageLoader to OldImageLoader
COAND-673
