+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/giftcard/GiftCardPaymentConfirmationDialogFragment.kt
@@ -14,7 +14,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.adyen.checkout.components.api.ImageLoader
+import com.adyen.checkout.components.api.OldImageLoader
 import com.adyen.checkout.components.util.CurrencyUtils
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
@@ -95,7 +95,7 @@ internal class GiftCardPaymentConfirmationDialogFragment : DropInBottomSheetDial
 
         val paymentMethods = alreadyPaidMethods + currentPaymentMethod
 
-        val imageLoader = ImageLoader.getInstance(
+        val imageLoader = OldImageLoader.getInstance(
             requireContext(),
             dropInViewModel.dropInConfiguration.environment
         )
