+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/paymentmethods/PaymentMethodListDialogFragment.kt
@@ -19,7 +19,7 @@ import androidx.core.view.children
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
-import com.adyen.checkout.components.api.ImageLoader
+import com.adyen.checkout.components.api.OldImageLoader
 import com.adyen.checkout.components.model.paymentmethods.StoredPaymentMethod
 import com.adyen.checkout.components.ui.view.AdyenSwipeToRevealLayout
 import com.adyen.checkout.core.log.LogUtil
@@ -77,7 +77,7 @@ internal class PaymentMethodListDialogFragment :
     }
 
     private fun initPaymentMethodsRecyclerView() {
-        val imageLoader = ImageLoader.getInstance(
+        val imageLoader = OldImageLoader.getInstance(
             requireContext(),
             dropInViewModel.dropInConfiguration.environment
         )
