+++ paybybank/src/main/java/com/adyen/checkout/paybybank/PayByBankRecyclerAdapter.kt
@@ -13,13 +13,13 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.adyen.checkout.components.api.ImageLoader
+import com.adyen.checkout.components.api.OldImageLoader
 import com.adyen.checkout.components.ui.databinding.RecyclerListWithImageBinding
 import com.adyen.checkout.issuerlist.IssuerModel
 import com.adyen.checkout.issuerlist.R
 
 internal class PayByBankRecyclerAdapter(
-    private val imageLoader: ImageLoader,
+    private val imageLoader: OldImageLoader,
     private val paymentMethod: String,
     private val onItemClicked: (IssuerModel) -> Unit,
 ) : ListAdapter<IssuerModel, PayByBankRecyclerAdapter.PayByBankViewHolder>(IssuerDiffCallBack) {
@@ -40,7 +40,7 @@ internal class PayByBankRecyclerAdapter(
         fun bind(
             paymentMethod: String,
             issuerModel: IssuerModel,
-            imageLoader: ImageLoader,
+            imageLoader: OldImageLoader,
             onItemClicked: (IssuerModel) -> Unit,
         ) {
             binding.root.setOnClickListener { onItemClicked(issuerModel) }
