+++ voucher/src/main/java/com/adyen/checkout/voucher/VoucherView.kt
@@ -15,7 +15,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.widget.LinearLayout
 import androidx.browser.customtabs.CustomTabsIntent
-import com.adyen.checkout.components.api.ImageLoader
+import com.adyen.checkout.components.api.OldImageLoader
 import com.adyen.checkout.components.api.LogoApi
 import com.adyen.checkout.components.base.ComponentDelegate
 import com.adyen.checkout.components.extensions.setLocalizedTextFromStyle
@@ -42,7 +42,7 @@ internal class VoucherView @JvmOverloads constructor(
 
     private val binding: VoucherViewBinding = VoucherViewBinding.inflate(LayoutInflater.from(context), this)
 
-    private lateinit var imageLoader: ImageLoader
+    private lateinit var imageLoader: OldImageLoader
 
     private lateinit var localizedContext: Context
 
@@ -58,7 +58,7 @@ internal class VoucherView @JvmOverloads constructor(
         this.localizedContext = localizedContext
         initLocalizedStrings(localizedContext)
 
-        imageLoader = ImageLoader.getInstance(context, delegate.componentParams.environment)
+        imageLoader = OldImageLoader.getInstance(context, delegate.componentParams.environment)
 
         observeDelegate(delegate, coroutineScope)
